+++ docs/changes.rst
@@ -14,6 +14,7 @@ Released on ? 2016.
 * Added automatic fix to remove control chars.
 * Added per language overview to project.
 * Added support for CSV export.
+* Added CSV download for stats.
 
 weblate 2.9
 -----------